cut cache_entry refcounts to 32 bits

+++ src/vmemcache.h
@@ -78,7 +78,7 @@ struct vmemcache {
 
 struct cache_entry {
 	struct value {
-		uint64_t refcount;
+		uint32_t refcount;
 		int evicting;
 		struct repl_p_entry *p_entry;
 		size_t vsize;
